backend/requirements.txt
fastapi
uvicorn
pydantic
requests

Pillow
pillow-heif